Some cleanup

+++ src/code/code_800FC620.c
@@ -20,11 +20,11 @@ char D_80134488[0x18] = {
     0xFF, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x00,
 };
 
-s32 Overlay_Load(u32 vRomStart, u32 vRomEnd, void* vRamStart, void* vRamEnd, void* allocatedVRamAddr) {
+s32 Overlay_Load(uintptr_t vRomStart, uintptr_t vRomEnd, void* vRamStart, void* vRamEnd, void* allocatedVRamAddr) {
     s32 pad[3];
     uintptr_t end;
     OverlayRelocationSection* ovl;
-    uintptr_t ovlOffset;
+    u32 ovlOffset;
     size_t size;
 
     size = vRomEnd - vRomStart;
